backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
langchain==0.3.20
langchain-community==0.3.18
langchain-core==0.3.45
chromadb==0.6.3
replicate==1.0.4
sentence-transformers==3.4.1
langchain-huggingface==0.1.2
gunicorn==23.0.0
